Add a startup pause for new EC2 servers. Related to #244.
Like DigitalOcean, new EC2 servers occasionally appear to be responsive
(e.g. the SSH port is open, and the daemon is running) but they aren't
ready to handle incoming connections in a timely fashion. This commit
adds a 90-second delay (which can be bypassed), giving new servers some
breathing room to get it together.