Prevent hanging on re-run
Running this script on certain servers fails a lot of times and requires 4-5 re-executions of ansible-playbook. I'm not trying to fix that.

After one of the re-runs Nginx was already running but client-side docs weren't generated yet, and Ansible hanged on "Make sure that nothing is listening on port 80". Stopping nginx manually allowed Ansible to proceed, so my commit is making this automated.