Create option to use FQDN for streisand server certificates
The default setting in Streisand is to create server certficates
based on the IP address. This creates error messages when the IP
changes.

The IP might change for example on an AWS EC2 instance
that is stopped and restarted later, or moved to another region.
To avoid the hassle of re-creating and distributing certificates,
a server should be able to have a DNS entry (ie, vpn.example.com).

This edit creates this option. A new variable
   streisand_cn_name
can be defined in
   playbooks/roles/common/vars/main.yml

If it is defined, the CN name of the certificates will change to
whatever domain name streisand_cn_name is set to.

If it is not defined, there is no change in behaviour.

CAUTION:
* Only tested on openconnect and AWS.
* The documentation might also need adaptation (currently a mix of IP and hostname.
* Did not touch TOR.