Move the Python SNI tasks back to the L2TP/IPsec role.
Temporary fix for an SNI error that was occurring when downloading the
OpenConnect source code from the CloudFront CDN when these supporting
packages were in place. SNI support in Python is weird.

The real fix will be to move to a more modern version of Python that
natively supports SNI in Ubuntu 16.04.