Use environment variable as default for DO token
Default the `do_access_token` to the environment variable
`DIGITALOCEAN_API_KEY`. This variable is used by doctl and other
tools and is often exported in the environment of a user which
leverages these tools.

The result is a little bit ugly in the prompt format, but it does
make provisioning easier from a user perspective:

> If you enter nothing here, the environment variable
> DIGITALOCEAN_API_KEY will be used.
>
> What is your DigitalOcean Access Token?
> [{{ lookup('env', 'DIGITALOCEAN_API_KEY') }}]: