[AWS] Add autorecover alarm
This task configures a CloudWatch metric alarm to trigger auto-recovery
if the host system is down for 2 minutes.

AWS supports a feature known as auto-recovery via CloudWatch, which
will respawn an instance on host system failure. The recovered instance
is identical to the original instance, including instance ID,
private/elastic IP addresses, and metadata.

Currently this feature is only available in the `us-east-1` region for
the C3, C4, M3, R3, and T2 instance types.