Create and reference streisand-users mailing list
There is a lot of chatter within the Github issues for things like
troubleshooting connectivity and basic installation questions which
would be better suited for a mailing list.

I took the liberty of creating a list on Google Groups and adding it
to the documentation here. The idea is that this will reduce clutter
in the Github issues (which should arguably be only code-related issues)
and provide users with a way to discuss non-code issues.