Updates shadowsocks-libev to 3.0.5, modern config.
This commit updates the shadowsocks role to install & configure
shadowsocks-libev 3.0.5 (the current release at time of writing) instead
of 2.5.6. In the process of doing so the deprecated OTA config option is
removed and the configured ciphersuite is switched from aes-256-cfb to
chacha20-ietf-poly1305.

Prior to this commit the build process involved installing a number of
dependencies that were only required for documentation or for building
a .deb to install. In the interest of slimming down the role this commit
opts to install the minimum required dependencies and installs directly
from source. The role templates its own systemd defaults and copies the
systemd service definition that the deb _would_ have installed directly
onto the system. This is a net savings of around ~800mb of dependencies
but means that there is a) no man file b) uninstallation must be done
with the Makefile instead of apt-get/dpkg.

This commit **does not** update the client documentation. This commit
does delete all references to OTA mode & removes the associated template
variable. There should be another pass done on the docs to make sure the
client recommendations are sound. Similarly, this commit does not update
the clients that are mirrored to the gateway and a pass should be done
to modernize those.