Ensures shadowsocks-libev is listening.
This commit adds a `wait_for` task to the shadowsocks role to ensure
that `shadowsocks_server_port` on the `streisand_ipv4_address` is in
a listening state before the role proceeds. If the preceding `systemd`
task worked as we expect then there is a `ss-server` process bound to
the server port.