Updates shadowsocks client mirror & docs.
This commit updates the shadowsocks client mirror to match the
shadowsocks-libev update.

* The Android client is updated from 2.10.7 to **4.1.4**
* The Windows client is updated from 3.0 **4.0.1**
* The OSX client is updated from 1.2 to **1.5-beta.2**
* The Linux python client is replaced with **go-shadowsocks2 0.0.8*

Using a beta release of the OSX client is the only way to get AEAD
support without replacing the client with an alternative choice.

For Linux clients we previously recommended the Python pip `shadowsocks`
project, which remains at 2.8.2 (the version we are were mirroring
previous to this commit) and appears unmaintained. Since there's
a nicely maintained Go client with a responsive author this commit
replaces the Linux client mirror to use the x64 binary release of the
`go-shadowsocks2` project. One day I would like to see the
`shadowsocks-libev` server side role updated to use `go-shadowsocks2` as
well and so this is also a step towards that direction. The shadowsocks
Linux instructions are updated to use the new client.

The iOS shadowsocks instructions are updated to explicitly mention that
Potatso **does not support AEAD** and can't be used with the current
configuration. After some testing it appears only Cross on iOS supports
AEAD and so we recommend it. Shadowrocket is another option but is
currently only supporting AEAD in a limited test release.