Adds workaround for Linode's missing kernel headers.
Building on https://github.com/jlund/streisand/pull/612 this commit adds
a workaround for Linode's grub2 instances not having kernel headers
installed. This manifests as the `wireguard-dkms` package installing but
generating no kernel modules. Later the systemctl task will fail to
bring the wg0-server interface up as a result.

This commit works around this by doing the equivalent of:
  `apt-get install linux-kernel-headers-$(uname -r)`
on the newly provisioned server, ensuring the kernel headers are there
for the wireguard role.