Restores `ansible_python_interpreter` inv var.
In 89f40e8 we revoved the `ansible_python_interpreter` on the assumption
it broke for many distros unless modified. This is likely still true,
but removing it also breaks OSX hosts that were previously working.

This commit restores the `ansible_python_interpreter` var until a less
jarring fix can be worked out.