Allows host var to skip gateway pause & prompt.
This commit moves the streisand-gateway tasks related to pausing
execution, prompting for continuation, and opening the generated docs in
the host browser into an ansible block. The block has a condition
attached such that it only executes when the `noninteractive` var is
anything other than `"true"`.

The Vagrantfile is updated to provide `noninteractive = "true"` as a
host inventory variable to exploit this change. Now its possible to
`vagrant up` and have the command return without needing to use
a keydown to continue through the pause.