Modern Shadowsocks server, config, and client options (#611)
* Updates shadowsocks-libev to 3.0.5, modern config.

This commit updates the shadowsocks role to install & configure
shadowsocks-libev 3.0.5 (the current release at time of writing) instead
of 2.5.6. In the process of doing so the deprecated OTA config option is
removed and the configured ciphersuite is switched from aes-256-cfb to
chacha20-ietf-poly1305.

Prior to this commit the build process involved installing a number of
dependencies that were only required for documentation or for building
a .deb to install. In the interest of slimming down the role this commit
opts to install the minimum required dependencies and installs directly
from source. The role templates its own systemd defaults and copies the
systemd service definition that the deb _would_ have installed directly
onto the system. This is a net savings of around ~800mb of dependencies
but means that there is a) no man file b) uninstallation must be done
with the Makefile instead of apt-get/dpkg.

This commit **does not** update the client documentation. This commit
does delete all references to OTA mode & removes the associated template
variable. There should be another pass done on the docs to make sure the
client recommendations are sound. Similarly, this commit does not update
the clients that are mirrored to the gateway and a pass should be done
to modernize those.

* Fixes shadowsocks `rng-tools` dep package name.

This commit fixes the shadowsocks role's `shadowsocks_dependencies`
variable to use the correct name for the `rng-tools` package. Previously
it was listed as `rngtools` causing an apt failure.

* Removes `-auth` suffix from shadowsocks QR code auth method.

This commit removes the `-auth` suffix that was being appended to the
`{{ shadowsocks_encryption_method }}` var when constructing the
shadowsocks QR code URI. This was previously required to signify OTA
mode, which has been removed/deprecated.

* Removes --no-install-recommends TODO

* Ensures shadowsocks-libev is listening.

This commit adds a `wait_for` task to the shadowsocks role to ensure
that `shadowsocks_server_port` on the `streisand_ipv4_address` is in
a listening state before the role proceeds. If the preceding `systemd`
task worked as we expect then there is a `ss-server` process bound to
the server port.

* Updates shadowsocks client mirror & docs.

This commit updates the shadowsocks client mirror to match the
shadowsocks-libev update.

* The Android client is updated from 2.10.7 to **4.1.4**
* The Windows client is updated from 3.0 **4.0.1**
* The OSX client is updated from 1.2 to **1.5-beta.2**
* The Linux python client is replaced with **go-shadowsocks2 0.0.8*

Using a beta release of the OSX client is the only way to get AEAD
support without replacing the client with an alternative choice.

For Linux clients we previously recommended the Python pip `shadowsocks`
project, which remains at 2.8.2 (the version we are were mirroring
previous to this commit) and appears unmaintained. Since there's
a nicely maintained Go client with a responsive author this commit
replaces the Linux client mirror to use the x64 binary release of the
`go-shadowsocks2` project. One day I would like to see the
`shadowsocks-libev` server side role updated to use `go-shadowsocks2` as
well and so this is also a step towards that direction. The shadowsocks
Linux instructions are updated to use the new client.

The iOS shadowsocks instructions are updated to explicitly mention that
Potatso **does not support AEAD** and can't be used with the current
configuration. After some testing it appears only Cross on iOS supports
AEAD and so we recommend it. Shadowrocket is another option but is
currently only supporting AEAD in a limited test release.

* Ensures shadowsocks config file is readable by nogroup.

* Fixes shadowsocks client mirroring.

Fixes one mismatched checksum, two broken urls.