Vagrant dev env support & basic end-to-end service tests (#660)
This PR adds a `Vagrantfile` that will create & provision two VMs at fixed IP addresses: `streisand-host` and `streisand-client`.

The host machine is provisioned using the base Streisand playbooks with very little modification (primarily modified to allow a non-interactive promptless mode). One workaround is in place in `playbooks/vagrant.yml` to ensure that the correct interface is used for the `ansible_default_ipv4` var.

Along the way I also updated the SSH role to generate a known-hosts file with SSH fingerprints for the server. The fingerprints file is put in the webroot for clients to download instead of doing TOFU SSH fingerprint matching themselves.

The client machine is provisioned with a new `test-client` role. The goal of this role is to provision a machine that downloads resources from the Streisand host's HTTPS gateway and uses those resources to test the various services end-to-end.

Initially I have added client support for:
* testing parts of the gateway (HTTP auth, etc)
* testing the SSH forward user (SOCKS proxy, key auth, using the known hosts file) and
* testing Shadowsocks (installing the Go client from the host's mirror, connecting to the shadowsocks-libev server and proxying a request).

The tests are implemented as bash scripts instead of Ansible tasks largely because they involve setting up and tearing down proxies and the complexity merits having a more flexible environment (I think).