Updates tor-bridge role to use var for publish desc.
This commit updates the tor-bridge role to use a var
(`tor_publish_service_desc`) with a default value of `0` as the value
for the torrc.j2 template's `PublishServerDescriptor` config value. This
ensures that by default the tor bridge is private but it is easy to
change for those that might wish to publish their bridge.