Add support for local provisioning.
This commit adds initial first-party support for "local provisioning",
that is, installing Streisand to the same machine that is running the
Ansible playbooks. This feature can be useful for individuals that can't
meet the Ansible system requirements on their client machine (e.g.
because it is Windows), or for users whose internet connection makes the
Ansible SSH connections unreliable.

To support local provisioning a new option is added to the initial
./streisand script menu. It is marked as an advanced feature and
requires that you enter a confirmation prompt to ensure you fully
understand that the Streisand playbooks will be overwriting configs on
the same machine that is running the script.

The Streisand script is updated to pull distinct operations into bash
functions. A new test for the existence of the ~/.ssh/id_rsa.pub SSH key
is added. Finally the `shellcheck` utility is added to CI and run
against the `streisand` shell script as part of each CI build to ensure
we don't fall into any of the many BASH "gotchas".

Note: The `streisand_ip4_address` var is set to the Ansible default IPv4
address, which itself is calculated based on the default IPv4 route to
8.8.8.8. This works for _most_ systems but will likely fail for some
more esoteric providers. I suggest that we workaround those in
`localhost.yml` on a case-by-case basis as they occur.