Updates mirrored Shadowsocks Android APK to 4.1.8.
The `shadowsocks-android` project has tagged a `4.1.8` release today
that reintroduces the built-in QR code scanner. Some Streisand users
have found this feature helpful in the event they don't have the ZXing
QR code scanner and can't access Google Play to get it.

This commit updates the Streisand mirror's copy of the Shadowsocks
Android client to version 4.1.8.