Support GCE w/ local & existing provisioning.
This commit adds support for GCE instances for the "advanced mode" local
& existing server provisioning options.

To do so an attempt at provider detection is added that allows the
localhost.yml and existing-server.yml roles to act like a genesis role
and, when run against a GCE instance, set the streisand IPv4 address to
the external IP from metadata. This is required for GCE because none of
the machine's interfaces have the external IP.

In the process I adjusted the Shadowsocks role's test for the
shadowsocks-libev service to test both the external IPv4 address and
the internal IPv4 address from the Ansible default interface. This is
a little hackish but seems to work reliably enough to let this PR move
forward. A more holistic approach should be evaluated as a follow-up.