Update client test for Shadowsocks to use server QR code.
This commit updates the Shadowsocks client test so that it no longer
relies on being provided the Shadowsocks password OOB by the Ansible
playbooks. Instead the test client can download the Shadowsocks QR code
from the Streisand server gateway instance & with the help of a little
Python script parse the values required. This makes remote testing
simpler (one less input to provide) and has the added benefit of also
testing the properties of the QR code and acting more like a _real_
client.