Modular deployment: Shadowsocks & POC
This commit introduces a proof-of-concept modular deployment of the
Shadowsocks role. This serves to set an example for how we can approach
the larger task of making the major Streisand services modular for
a pick-and-choose deployment. Shadowsocks was selected as a first role
since I'm familiar with it now and it's fairly small.

By default the `streisand_shadowsocks_enabled` var is set to `yes` to
include Shadowsocks. Once more of the roles are modularized a prompt
system will be introduced to toggle this on/off and the default value
can be revisited.