Add support for local provisioning. (#724)
This PR adds initial first-party support for "local provisioning",
that is, installing Streisand to the same machine that is running the
Ansible playbooks. This feature can be useful for individuals that can't
meet the Ansible system requirements on their client machine (e.g.
because it is Windows), or for users whose internet connection makes the
Ansible SSH connections unreliable.

The Streisand script is updated to pull distinct operations into bash
functions. A new test for the existence of the `~/.ssh/id_rsa.pub` SSH key
is added, since presently Streisand assumes it to exists and this is 
frequently a source of user error. 

To support local provisioning a new option is added to the initial
./streisand script menu. It is marked as an advanced feature and
requires that you enter a confirmation prompt to ensure you fully
understand that the Streisand playbooks will be overwriting configs on
the same machine that is running the script.

Similarly, an advanced option for running the Streisand playbook against
an existing server was added to the wrapper script. This allows testing the
SSH access for the root user (overidable with the `ANSIBLE_SSH_USER`
env var) before invoking the playbook, hopefully again helping to identify
user error early. The existing server option generates an inventory file on
the fly based on an IP the user enters and writes it to 
`inventories/inventory-existing`, avoiding the need to manually edit an
inventory file for existing servers & hopefully encouraging more people to
use the wrapper where we can detect common errors.

With several inventory files to `sed` patch for `python` vs `python2` and 
to avoid clutter this PR also moves the inventory files into a dedicated 
`inventories/` directory. The `ansible.cfg` file is updated to use 
`inventories/inventory` as the default inventory to maintain backwards
compatibility.

To help keep a growing pile of bash manageable `shellcheck` is added
to CI builds and run against the `streisand` wrapper.

For localhost installs the `streisand_ip4_address` var is set to
the Ansible default IPv4 address, which itself is calculated based on
the default IPv4 route to 8.8.8.8. This works for _most_ systems but
will likely fail for some more esoteric providers.

For the more complicated cases (e.g. GCE) a common 
`playbooks/provider-detect.yml` attempts to determine if the instance 
belongs to a known cloud provider (currently only GCE & AWS are 
supported) and will update the external IPv4 address according to the instance
metadata since it is not available as a property of any of the interfaces
on the server.

Resolves #595, #468, #122