Ensure apt state is clean.
With `update_cache` on the lxd install apt module invocation a "failure
to update apt cache" was observed. Since this is CI its easiest to just
ensure a totally clean & up to date apt state after we add the PPA and
remove the `update_cache` directive on the apt install. (Thanks for the
tip Ali!)