Fix CI by working around Ansible `apt_repository` bug. (#760)
The Ansible `apt_repository` module at the time of writing is choking on
the ubuntu-lxc ppa repository because one of the mirrors is missing
binary-i386 builds. We don't actually **use** i386 builds because Travis
is x86_64 by default. We can work around this error by using the
`add-apt-repository` via the Ansible `command` module instead of doing
things the "right" way. This can be reverted when the upstream
`add-apt-repository` module is fixed.

With `update_cache` on the lxd install apt module invocation a "failure
to update apt cache" was observed. Since this is CI its easiest to just
ensure a totally clean & up to date apt state after we add the PPA and
remove the `update_cache` directive on the apt install. (Thanks for the
tip Ali!)