Refactor/Cleanup test/test.sh
This commit updates `tests/tests.sh` to have less duplication & to pass
a shellcheck invocation. The commonality between it and the base
`streisand` script was extracted to `util/ansible_check.sh` and included
in both places. To make shellcheck follow the includes we add a `-x` to
the `.travis.yml` invocation of shellcheck.