CI enhancements/shell script cleanup (#761)
* Ensure early build exit for failed scripts.

Per Travis docs the way to have the build exit if one script fails is
to only specify one script and rely on shell operators to short circuit
on non-zero exit. Kind of ugly... but ok Travis............ if you say
so..........................

* Refactor/Cleanup test/test.sh

This commit updates `tests/tests.sh` to have less duplication & to pass
a shellcheck invocation. The commonality between it and the base
`streisand` script was extracted to `util/ansible_check.sh` and included
in both places. To make shellcheck follow the includes we add a `-x` to
the `.travis.yml` invocation of shellcheck.

* Document CI

This commit adds some high level documentation about CI to
`documentation/testing.md` to hopefully make the testing more
accessible to external contributors.