Fix tor daemon & obfs4proxy apparmor profile.
There is an outstanding bug[0] with the Tor apparmor profile that
causes the tor daemon to fail to start when configured with a hidden
service directory.

To temporarily work-around this issue this commit adds a new
`/etc/apparmor.d/local/system_tor` override file that adds the missing
capability. This can be removed when upstream has fixed the problem.

Additionally, while debugging this issue I removed the abstractions
apparmor file that Streisand previously used to replace the distro
provided profile. We only need to change two lines so rather than keep
an entire profile up to date this commit adds a `replace` task to update
the two obfsproxy binaries to have the correct `ix` permissions. We
should look into fixing this in a better way in the future.

[0]: https://bugs.debian.org/cgi-bin/bugreport.cgi?bug=862993