Fix simple-obfs git clone local modifications conflict. (#789)
With Shadowsocks simple-obfs v0.0.3 multiple invocations of the
Shadowsocks playbook will error with the following:

    "changed": false, "failed": true, "msg": "Local modifications exist
    in repository (force=no)."

Looking in the `simple_obfs_compilation_directory` we see that
`aclocal.m4` has local modifications:

```
/usr/local/src/shadowsocks-libev-0.0.3# git status
HEAD detached at v0.0.3
Changes not staged for commit:
  (use "git add <file>..." to update what will be committed)
  (use "git checkout -- <file>..." to discard changes in working directory)

	modified:   aclocal.m4
```

While `aclocal.m4` is present in the simple-obfs `.gitignore`, it was
checked into the repository beforehand causing this error after
autoconf mutates it from a build.

This has been fixed on the upstream simple-obfs master branch with
commit 8c22d275b29fc85e3db036c135f5d61f67585f20, but until there is
a new release tagged we can work around it by adding "force:yes" to the
git clone task and overwriting the repo even though there are changes.

In addition to adding the "force" directive this commit also updates the
`simple_obfs_compilation_directory`. Previously it was using the same
directory name as shadowsocks-libev, causing confusion.