Restore the unconditional ec2-security-group tasks.
As part of the modularization work the ec2-security-group task that
adds rules for each of the service ports was broken into three tasks:

1. a task for all of the unmodular ports.
2. a task that opened Shadowsocks ports when
   `streisand_shadowsocks_enabled` was true
3. a task that opened Wireguard ports when
   `streisand_wireguard_enabled was true.

This design assumed that multiple `ec2_group` invocations was additive,
that is, that the three tasks would add to the same security group.
Instead the last `ec2_group` task is the only one that is used, in
effect only opening the Wireguard port for a default install.

I suspect there is a way to properly open only the required ports based
on the `enabled` vars but in the interest of fixing a regression on
master quickly this commit restores the unconditional EC2 security group
role with one `ec2_group` task.

This has the result of opening ports for Wireguard & Shadowsocks on EC2
when those services aren't installed. This isn't _ideal_, but since we
already ship `ufw` there isn't significant security impact since the
host machine will drop traffic to these ports regardless

I will investigate better options this weekend. Apologies for the
regression!