Set GCE `ansible_user` to `root` to match SSH key metadata.
Prior to this commit the GCE genesis role was setting the `ansible_user`
to `ubuntu`. This appears to be a mistake since in the earlier `gce`
task it adds the SSH key under the `root` user.

In testing a GCE provision recently I noticed that the cloud-init task
would fail with an SSH "Permission denied (publickey)" error. Similar
`raw` tasks would fail if the cloud-init include was commented out.

In practice using `root` as the `ansible_user` for GCE allows
provisioning to complete without error.

It's not clear to me why this didn't break before since the
`ansible_user` has been static for some time... Mysterious!