Modular Streisand: Optional Tor (#808)
This commit updates the Tor role to be optional, allowing it to be enabled/disabled by changing the `playbooks/group_vars/all` variable `streisand_tor_enabled`. By default it is true and the base Streisand system remains unchanged.

This largely follows in the footsteps of #794 and #802 and doesn't have much of note. The only "tricky" bit was removing the Nginx hidden service vhost from the `streisand-gateway` role. It is now a separate vhost templated into `/etc/nginx/sites-available` (and `sites-enabled`) under the name `streisand-hidden-service`. Now it is only created & enabled when the Tor role is applied.

Updates #746