Optional SSH forward role.
This commit splits the SSH forward user from the base `ssh` role into
a separate `ssh-forward` role. This new forward role is conditionally
included in a Streisand build based on the value of
`streisand_ssh_forward_enabled` (which defaults to true to maintain the
existing "full" Streisand build).

Notably the firewall configs are *not* modified because the base SSH
daemon will be installed regardless of whether
`streisand_ssh_forward_enabled` as part of the base system. E.g. port 22
will always be open.