Update the Monit config to use the `service` init wrapper. (#874)
When the handler to restart Nginx was skipped due to #873, Monit would
start the Nginx daemon using the legacy `/etc/init.d/nginx start`
command. This broke future invocations of the `service nginx restart`
and `service nginx stop` commands. The same was true for other daemons
that Monit launched (or re-launched).

@alimakki first documented the symptoms of this problem in #871.
Although default installations were insulated from this bug, and #873
should prevent it from occurring again in the future, Monit can still
be a good init citizen until we replace it with something better.