ssh-forward: Allow port forwarding only
By default one can log into the Streisand server with the SSH private
key, which is intended for port forwarding. This poses a security issue
if a careless sysadmin sets open permissions on critical files.
Prepending a force command to authorized_key is a simple way to
prevent SSH login with that key.