ssh-forward: Allow port forwarding only (#883)
By default one can log into the Streisand server with the SSH private
key, which is intended for port forwarding. This poses a security issue
if a careless sysadmin sets open permissions on critical files.
This commit prepends a force command to authorized_key as a 
simple way to prevent SSH login with the SSH forward user key.

Removing support for a shell from this user means sshuttle is no longer
functional. At this point it seems like the use-case for sshuttle has been
outmoded and so this commit also removes sshuttle from Streisand.

Documentation is updated to reflect the requirement to change the SSH
client configuration to only forward traffic, not create an interactive shell
session.