Collect diagnostics information locally to share with new issues.
This commit adds a diagnostics role that is used to collect information
about the Ansible installation & the Streisand clone being run. This
information is written to `streisand-diagnostics.md` in the project
directory but **not shared anywhere**. Care was taken to not include
anything that is especially sensitive (e.g. IP addresses, BIOS info).

The information written to the `streisand-diagnostics.md` file includes:
* Ansible version
* Host OS/version
* Python version
* Streisand git revision
* Whether there are untracked changes in the git clone (e.g. you've
  mucked about!)
* The genesis role that was being used (Linode, EC2, Localhost, etc)
* Which optional roles were enabled
* How many VPN clients were being configured

This commit updates the Github issue template to indicate that users
opening new issues about Streisand should share the
`streisand-diagnostics.md` file contents. This should make
troubleshooting much easier for everyone!

To be 1000% clear: no information is collected and sent to anyone
without the user doing it themselves as part of a Github issue.