Provide alert threshold defaults to `linode` task.
Based on a comment[0] from Issue #626 this commit adds
`alert_x_threshold` values to the `linode` task in the `genesis-linode`
playbook. On the systems I've tested this is not necessary but also
harmless. For those affected by issue #626 it is necessary to avoid
a failure to create Linode servers based on the threshold values being
outside of validation range or undefined.

Thanks to @cmeid for suggesting the workaround and @i-s-o-g-r-a-m c for
testing the fix.

[0] - https://github.com/jlund/streisand/issues/626#issuecomment-319812261