Use defaults when accessing hostvars for diagnostics.
This commit updates the `hostvar`'s referenced in the
`streisand-diagnostics md.j2` template to assume nothing about which
vars are/aren't present on the system. For mysterious reasons it seems
like sometimes the same version of Ansible will have different hostvars
populated. Since we intend for this diagnostics information to be useful
in cases when things are already going wrong we should take the most
conservative approach possible in collecting the information.

This commit updates the template to use the `.get()` style of accessing
Python dictionary keys, providing a default value in the event the key
isn't present. This should allow for the template to render even when
none of the referenced Ansible hostvars are available.