Conditionally accesses hostvar keys for diagnostics.
This commit updates the `streisand-diagnostics md.j2` template to
conditionally reference hostvars for diagnostics. If the hostvars don't
exist the diagnostic is templated as Unknown.

For mysterious reasons it seems like sometimes the same version of
Ansible will have different hostvars populated. Since we intend for this
diagnostics information to be useful in cases when things are already
going wrong we should take the most conservative approach possible in
collecting the information.

One might be tempted to use `.get('foo', 'Unknown')` or similar instead
of a mess of `if/else/endif` blocks. Unfortunately the Ansible
`hostvars` var isn't a normal dict and this will still fail if `'foo'`
isn't defined.