New role: Let's Encrypt issued nginx certificate (#911)
This role implements Let's Encrypt support for Streisand servers with a public domain name. It is intended to resolve https://github.com/jlund/streisand/issues/639.

The role adds two more questions during playbook run. Whether the role is run is based on user input. If user skips the questions, the role will not be run at all.

The role itself has a fail-safe. If it successfully gets the certificates, it sets `le_ok` true, which has the following consequences:
- OpenSSL self-signed certificates generation process will be skipped;
- acmetool (https://github.com/hlandau/acme) is installed on the server 
- acmetool is used to provision a Let's Encrypt issued certificate that will be used in Nginx, instead of self-signed ones;
- Instructions on how to import self-signed certificates in generated_docs will be removed.

If the roles fails to get the certificates for any reason (Let's Encrypt server down, user error, etc), `le_ok` would be set false. The consequences above will not happen, the playbook continues to run as if the role does not exist at all.

Acmetool handles installing a cron task to renew the certificate as required in an automated fashion.