Support specifying custom SSH public key.
Prior to this commit Streisand universally assumed the user's SSH
public key was `~/.ssh/id_rsa.pub`. This SSH key was then used to
provision cloud instances.

This commit adds a new `streisand_ssh_key` var to the
`global_vars/default-site.yml` file. The site-specific override for
these defaults, `~/.streisand/site.yml` is now customized by
`playbooks/customize.yml` when the user specifies that they want to use
a different SSH pubkey. The user is expected to have the private
component loaded into their ssh agent or to customize Ansible to know
how to load it.

The validation playbook is updated to check that the `streisand_ssh_key`
exists. The corresponding shell based checks for the SSH key were
removed from the `./streisand` wrapper script since they are now
replaced by these Ansible tasks. Similarly the existing server playbook
now performs its own quick test of a raw_cmd on the server instead of
doing a SSH check in the `./streisand` wrapper. This means things fail
a little bit later in the process if you can't actually SSH to the
existing server but it isn't *much* later in the process and its nice to
rm some shell script code.

Three unrelated bugs were fixed with the existing server provisioning.
First: There was a typo in the `./streisand` wrapper script's
`local_provision` function that was running a non-existent playbook.
Second: the genesis role was not being saved as a fact on localhost
causing the diagnostics to print "unknown" for the genesis role used.
Lastly, because no `gather_facts: yes` for the localhost was run for the
existing server provision other diagnostic information was missing. This
was also the root cause of Issue #739. All of the above are now fixed!
Yay \o/